+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * When using the Projects Plugin, generated Features now get the same restricted status and Projects as the Location they are generated from.
 * [HELM] Fix livenessProbe for postgresql to use customized username with pg_isready.
 * Fixed double-initialisation under heavy load.
+* Added tests for MQTT-Create and fine-grained authorisation.
 
 
 ## Release version 2.5.2